requirements.txt
boto3
requests
python-dotenv
localstack
awscli-local
